+++ projects/library-ts/src/hello-world.ts
@@ -1,0 +1,7 @@
+export function sayHello(name: string) {
+	console.log(`Hi ${name}`);
+}
+
+export function sayGoodbye(name: string) {
+	console.log(`Goodbye ${name}`);
+}