+++ packages/toolkit/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file, per [the Keep a Changelog standard](http://keepachangelog.com/).
 
+## 1.0.3
+- Update @10up/eslint-config to 2.3.5
+- Update @10up/babel-preset-defaylt to 1.1.2
+- Add TypeScript support
+
 ## 1.0.2
 - Update @10up/stylelint-config to 1.1.1